+++ src/Oro/Bundle/FrontendBundle/Tests/Unit/Provider/FileIconProviderTest.php
@@ -4,42 +4,39 @@ namespace Oro\Bundle\FrontendBundle\Tests\Unit\Provider;
 
 use Oro\Bundle\AttachmentBundle\Entity\File;
 use Oro\Bundle\AttachmentBundle\Provider\FileIconProvider as BaseFileIconProvider;
-use Oro\Bundle\CacheBundle\Adapter\ChainAdapter;
 use Oro\Bundle\FrontendBundle\Provider\FileIconProvider;
 use Oro\Bundle\FrontendBundle\Request\FrontendHelper;
 use Oro\Component\Layout\Extension\Theme\Model\CurrentThemeProvider;
-use Oro\Component\Layout\Extension\Theme\Model\Theme;
 use Oro\Component\Layout\Extension\Theme\Model\ThemeManager;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Psr\Cache\CacheItemPoolInterface;
-use Symfony\Contracts\Cache\CacheInterface;
-use Symfony\Contracts\Cache\ItemInterface;
 
-class FileIconProviderTest extends TestCase
+final class FileIconProviderTest extends TestCase
 {
     // Storefront icons
-    private const JPEG_ICON = 'jpeg-icon';
-    private const MP3_ICON = 'jpeg-icon';
-    private const DEFAULT_ICON = 'default-icon';
-    private const JPEG_EXT = 'jpeg';
-    private const MP3_EXT = 'mp3';
-    private const DEFAULT_EXT = 'default';
-    private const FILE_ICONS = [self::JPEG_EXT => self::JPEG_ICON, self::DEFAULT_EXT => self::DEFAULT_ICON];
-    private const FALLBACK_ICON = 'add-note';
+    private const string JPEG_ICON = 'jpeg-icon';
+    private const string MP3_ICON = 'jpeg-icon';
+    private const string DEFAULT_ICON = 'default-icon';
+    private const string JPEG_EXT = 'jpeg';
+    private const string MP3_EXT = 'mp3';
+    private const string DEFAULT_EXT = 'default';
+    private const array FILE_ICONS = [
+        self::JPEG_EXT => self::JPEG_ICON,
+        self::DEFAULT_EXT => self::DEFAULT_ICON
+    ];
+    private const string FALLBACK_ICON = 'add-note';
 
     // Backend icons
-    private const JPEG_FA_ICON = 'jpeg-fa-icon';
-    private const DEFAULT_FA_ICON = 'default-fa-icon';
-    private const FA_FILE_ICONS = [self::JPEG_EXT => self::JPEG_FA_ICON, self::DEFAULT_EXT => self::DEFAULT_FA_ICON];
+    private const string JPEG_FA_ICON = 'jpeg-fa-icon';
+    private const string DEFAULT_FA_ICON = 'default-fa-icon';
+    private const array FA_FILE_ICONS = [
+        self::JPEG_EXT => self::JPEG_FA_ICON,
+        self::DEFAULT_EXT => self::DEFAULT_FA_ICON
+    ];
 
-    private FrontendHelper|MockObject $frontendHelper;
-
-    private CurrentThemeProvider|MockObject $currentThemeProvider;
-
-    private ThemeManager|MockObject $themeManager;
-
-    private CacheInterface|CacheItemPoolInterface|MockObject $cache;
+    private FrontendHelper&MockObject $frontendHelper;
+    private CurrentThemeProvider&MockObject $currentThemeProvider;
+    private ThemeManager&MockObject $themeManager;
 
     private FileIconProvider $provider;
 
@@ -49,14 +46,12 @@ class FileIconProviderTest extends TestCase
         $this->frontendHelper = $this->createMock(FrontendHelper::class);
         $this->currentThemeProvider = $this->createMock(CurrentThemeProvider::class);
         $this->themeManager = $this->createMock(ThemeManager::class);
-        $this->cache = $this->createMock(ChainAdapter::class);
 
         $this->provider = new FileIconProvider(
             new BaseFileIconProvider(self::FA_FILE_ICONS),
             $this->frontendHelper,
             $this->currentThemeProvider,
-            $this->themeManager,
-            $this->cache
+            $this->themeManager
         );
     }
 
@@ -72,7 +67,7 @@ class FileIconProviderTest extends TestCase
         $file = new File();
         $file->setExtension($fileExtension);
 
-        self::assertEquals($expectedIcon, $this->provider->getExtensionIconClass($file));
+        self::assertSame($expectedIcon, $this->provider->getExtensionIconClass($file));
     }
 
     public function getExtensionIconClassBackendDataProvider(): array
@@ -105,13 +100,13 @@ class FileIconProviderTest extends TestCase
 
         $file = (new File())->setExtension(self::DEFAULT_EXT);
 
-        self::assertEquals(self::FALLBACK_ICON, $this->provider->getExtensionIconClass($file));
+        self::assertSame(self::FALLBACK_ICON, $this->provider->getExtensionIconClass($file));
     }
 
     /**
      * @dataProvider getExtensionIconClassStorefrontDataProvider
      */
-    public function testGetExtensionIconClassStorefrontNoCache(string $fileExtension, string $expectedIcon): void
+    public function testGetExtensionIconClassStorefront(string $fileExtension, string $expectedIcon): void
     {
         $this->frontendHelper->expects(self::exactly(2))
             ->method('isFrontendRequest')
@@ -121,61 +116,16 @@ class FileIconProviderTest extends TestCase
         $this->currentThemeProvider->expects(self::once())
             ->method('getCurrentThemeId')
             ->willReturn($themeName);
-
-        $this->cache->expects(self::once())
-            ->method('get')
-            ->with('oro_frontend.provider.file_icons_mapping.theme.' . $themeName)
-            ->willReturnCallback(function ($cacheKey, $callback) {
-                $item = $this->createMock(ItemInterface::class);
-
-                return $callback($item);
-            });
-
-        $themeHierarchy = [
-            $this->getTheme(['file_icons' => self::FILE_ICONS]),
-            $this->getTheme([]),
-            $this->getTheme(
-                ['file_icons' => [self::DEFAULT_EXT => self::DEFAULT_ICON, self::MP3_EXT => self::MP3_ICON]]
-            ),
-        ];
 
         $this->themeManager->expects(self::once())
-            ->method('getThemesHierarchy')
-            ->willReturn($themeHierarchy);
+            ->method('getThemeConfigOption')
+            ->with($themeName, 'icons')
+            ->willReturn(['file_icons' => [$fileExtension => $expectedIcon]]);
 
         $file = new File();
         $file->setExtension($fileExtension);
 
-        self::assertEquals($expectedIcon, $this->provider->getExtensionIconClass($file));
-    }
-
-    /**
-     * @dataProvider getExtensionIconClassStorefrontDataProvider
-     */
-    public function testGetExtensionIconClassStorefrontFromCache(string $fileExtension, string $expectedIcon): void
-    {
-        $fileIcons = array_merge(self::FILE_ICONS, [self::MP3_EXT => self::MP3_ICON]);
-        $this->frontendHelper->expects(self::exactly(2))
-            ->method('isFrontendRequest')
-            ->willReturn(true);
-
-        $themeName = 'default';
-        $this->currentThemeProvider->expects(self::once())
-            ->method('getCurrentThemeId')
-            ->willReturn($themeName);
-
-        $this->cache->expects(self::once())
-            ->method('get')
-            ->with('oro_frontend.provider.file_icons_mapping.theme.' . $themeName)
-            ->willReturn($fileIcons);
-
-        $this->themeManager->expects(self::never())
-            ->method(self::anything());
-
-        $file = new File();
-        $file->setExtension($fileExtension);
-
-        self::assertEquals($expectedIcon, $this->provider->getExtensionIconClass($file));
+        self::assertSame($expectedIcon, $this->provider->getExtensionIconClass($file));
     }
 
     public function getExtensionIconClassStorefrontDataProvider(): array
@@ -206,7 +156,7 @@ class FileIconProviderTest extends TestCase
             ->method('isFrontendRequest')
             ->willReturn(false);
 
-        self::assertEquals(self::FA_FILE_ICONS, $this->provider->getFileIcons());
+        self::assertSame(self::FA_FILE_ICONS, $this->provider->getFileIcons());
     }
 
     public function testGetFileIconsStorefrontNoTheme(): void
@@ -222,7 +172,7 @@ class FileIconProviderTest extends TestCase
         self::assertEmpty($this->provider->getFileIcons());
     }
 
-    public function testGetFileIconsStorefrontNoCache(): void
+    public function testGetFileIconsStorefront(): void
     {
         $this->frontendHelper->expects(self::once())
             ->method('isFrontendRequest')
@@ -233,64 +183,13 @@ class FileIconProviderTest extends TestCase
             ->method('getCurrentThemeId')
             ->willReturn($themeName);
 
-        $this->cache->expects(self::once())
-            ->method('get')
-            ->with('oro_frontend.provider.file_icons_mapping.theme.' . $themeName)
-            ->willReturnCallback(function ($cacheKey, $callback) {
-                $item = $this->createMock(ItemInterface::class);
-
-                return $callback($item);
-            });
-
-        $themeHierarchy = [
-            $this->getTheme(['file_icons' => self::FILE_ICONS]),
-            $this->getTheme([]),
-            $this->getTheme(
-                ['file_icons' => [self::DEFAULT_EXT => self::DEFAULT_ICON, self::MP3_EXT => self::MP3_ICON]]
-            ),
-        ];
+        $fileIcons = \array_merge(self::FILE_ICONS, [self::MP3_EXT => self::MP3_ICON]);
 
         $this->themeManager->expects(self::once())
-            ->method('getThemesHierarchy')
-            ->willReturn($themeHierarchy);
+            ->method('getThemeConfigOption')
+            ->with($themeName, 'icons')
+            ->willReturn(['file_icons' => $fileIcons]);
 
-        self::assertEquals(
-            array_merge(self::FILE_ICONS, [self::MP3_EXT => self::MP3_ICON]),
-            $this->provider->getFileIcons()
-        );
-    }
-
-    public function testGetFileIconsStorefrontFromCache(): void
-    {
-        $fileIcons = array_merge(self::FILE_ICONS, [self::MP3_EXT => self::MP3_ICON]);
-        $this->frontendHelper->expects(self::once())
-            ->method('isFrontendRequest')
-            ->willReturn(true);
-
-        $themeName = 'default';
-        $this->currentThemeProvider->expects(self::once())
-            ->method('getCurrentThemeId')
-            ->willReturn($themeName);
-
-        $this->cache->expects(self::once())
-            ->method('get')
-            ->with('oro_frontend.provider.file_icons_mapping.theme.' . $themeName)
-            ->willReturn($fileIcons);
-
-        $this->themeManager->expects(self::never())
-            ->method(self::anything());
-
-        self::assertEquals($fileIcons, $this->provider->getFileIcons());
-    }
-
-    private function getTheme(array $iconsConfig): Theme
-    {
-        $theme = $this->createMock(Theme::class);
-        $theme->expects(self::once())
-            ->method('getConfigByKey')
-            ->with('icons')
-            ->willReturn($iconsConfig);
-
-        return $theme;
+        self::assertSame($fileIcons, $this->provider->getFileIcons());
     }
 }